query-window: Perform complete re-iteration for query flow.
Add a new right permission -- UP_ApprovalNotRequired, so now on the databases where a user has this right -- allow him to execute the query after checking for the query type approval and a database selection -- without going into the approval flow of the query.

Case: II - After a query is set for approval -- now if the user changes the datbase-mapped through the dropdown or changes the raw Query written -- Automatically change the query status to HOLD_FOR_APPROVAL and save the query as Draft. User can now make the changes and again set the query for APPROVAL.

Case: III - After a query is approved -- user makes changes into it --> Change the status of query automatically to HOLD_FOR_APPROVAL and allow the changes to the query and user will have to set the query for approval again.

Case: IV - After a query has been executed/REJECTED -- User wants to execute the same query in some other database or make some changes to the rawQuery -- User has the option to `Make a Copy` of query -- which would redirect user to a new different query window with the details of the old query saved in the respective fields and user can perform operations on the query again treating it as a complete new query.